# Rio.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Rio.API",
    "version": "1.0"
  },
  "paths": {
...
    "/customRichText/{customRichTextTypeID}": {
      "get": {
        "tags": [
          "CustomRichText"
        ],
        "parameters": [
          {
            "name": "customRichTextTypeID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRichTextDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRichTextDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRichTextDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CustomRichText"
        ],
        "parameters": [
          {
            "name": "customRichTextTypeID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRichTextDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRichTextDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRichTextDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRichTextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRichTextDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRichTextDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRichTextDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/water-year-month/finalize": {
      "put": {
        "tags": [
          "WaterYearMonth"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WaterYearMonthDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaterYearMonthDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaterYearMonthDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CustomRichTextDto": {
        "type": "object",
        "properties": {
          "IsEmptyContent": {
            "type": "boolean"
          },
          "CustomRichTextID": {
            "type": "integer",
            "format": "int32"
          },
          "CustomRichTextType": {
            "$ref": "#/components/schemas/CustomRichTextTypeDto"
          },
          "CustomRichTextContent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomRichTextTypeDto": {
        "type": "object",
        "properties": {
          "CustomRichTextTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "CustomRichTextTypeName": {
            "type": "string",
            "nullable": true
          },
          "CustomRichTextTypeDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WaterYearDto": {
        "type": "object",
        "properties": {
          "WaterYearID": {
            "type": "integer",
            "format": "int32"
          },
          "Year": {
            "type": "integer",
            "format": "int32"
          },
          "ParcelLayerUpdateDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WaterYearMonthDto": {
        "type": "object",
        "properties": {
          "WaterYearMonthID": {
            "type": "integer",
            "format": "int32"
          },
          "WaterYear": {
            "$ref": "#/components/schemas/WaterYearDto"
          },
          "Month": {
            "type": "integer",
            "format": "int32"
          },
          "FinalizeDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}